x86, xen: use eagerfpu if CPUID shows task switch in TRAP_NM is unsafe
Currently we allow task switching during the device not available trap
if memory must be freed in order to allocate room for the FPU state,
on the assumption that TS is set and will generate a trap again if
the new task tries to use the FPU.  If TS is not set then the FPU state
might be clobbered by a new task.

According to the xen PVABI, the task switch bit in control register 0
will be cleared before entering the device not available trap.  This is not
compatible with the current trap handler.

For paravirt mode, unless the xen hypervisor indicates it will not
auto-clear the TS flag in CPUID Features 1, default to eagerfpu as eagerfpu
avoids the device not available trap entirely.

Reported-by: Zhu Yanhai <zhu.yanhai@gmail.com>
Signed-off-by: Sarah Newman <srn@prgmr.com>